refactor: added dbg messages to task dependency. Moved task dependency failures into task_dependency
